api/requirements.txt
cachelib==0.4.1
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
Deprecated==1.2.13
Flask==2.0.2
Flask-Session==0.4.0
gcloud==0.17.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
jwcrypto==1.0
jws==0.1.3
MarkupSafe==2.0.1
oauth2client==3.0.0
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
Pyrebase4==4.5.0
python-dotenv==0.19.1
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.1
